Added a transform component
Some javadoc and tests are still missing.
We also might want to take a look at the rotation member variable, which currently is a double (mostly due to the Math.PI-constant which too is a double). We might want to consider changing this to be a float, so that it requires less memory.