Added event-handling for the states
The statemanager listens to the states, who might send notifications on when to change state, or when to quit the game. The manager then takes the appropriate action.
The main activity listens to the statemanager. The game exits from the activity, so either the state manager has to notify the activity, or it needs a reference to it (which would cause a cyclic dependency). That's why the activity listens to the state manager.